Initial Commit
Not sure what I will be able to accomplish here, but I want to make my own vanilla implementation of a ZIP handling package, which is full ESM-compatible, uses the Compression Streams API for compression management, and provides it's own TS types.

There are already a bunch of JS libraries out there for this, but they seem to tend having a good amount of legacy things to manage, and maybe starting fresh could help make things simpler in the general scheme of things. That's what everyone says when they make their own library, but yeah.

So this will fill a similar void to what NBTify did for me, but for ZIP files as well.

https://en.wikipedia.org/wiki/ZIP_(file_format)